+++ prinz/src/main/scala/pl/touk/nussknacker/prinz/model/SignatureInterpreter.scala
@@ -1,0 +1,6 @@
+package pl.touk.nussknacker.prinz.model
+
+trait SignatureInterpreter {
+
+  def downloadSignature(modelId: String): Option[ModelSignature]
+}
